pyproject.toml
[tool.poetry]
name = "devind-django-dictionaries"
version = "0.5.0"
description = "Common dictionaries for devind applications"
authors = ["Victor <lyferov@yandex.ru>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.5"
requests = "^2.28.0"
beautifulsoup4 = "^4.11.1"
celery = "^5.2.7"
devind-helpers = "^1.0.0"
python-dotenv = "^0.20.0"
strawberry-graphql = {extras = ["debug-server"], version = "^0.114.6"}
strawberry-graphql-django = "^0.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
